Update and rename buttons-choices-task-modules.md to passing-parameters.md
While the original intention of this article was to describe buttons, pick lists and task modules, I believe it is much more helpful to users to describe the passing of parameters from user to recipe and recipe to recipe. That's because it was something that was very difficult to figure out when I first started learning about Workbot for MS Teams. While information about passing parameters was included here and there in the Workbot for MS Teams articles, they were all over the place and hard to understand. That is why I believe they should be consolidated into one article; descriptions of these Workbot for MS Teams features can be moved to other articles.